test: move shared code to new lib folder for reuse beyond chainsaw
