+++ test/e2e/chainsaw/tests/apis/members/memberWithoutRole/chainsaw-test.yaml
@@ -39,7 +39,7 @@ spec:
     - name: apiCrdFile
       value: (join('', ['resources/', $apiName, '.yaml']))
     - name: commandDir
-      value: "../../../../commands"
+      value: "../../../../../lib/commands"
     - name: gkoNamespace
       value: default
     - name: exportedApiYaml
