+++ test/e2e/chainsaw/tests/apis/members/removingMember/chainsaw-test.yaml
@@ -41,7 +41,7 @@ spec:
     # - name: apiCrdFile
     #   value: (join('', ['resources/', $apiName, '.yaml']))
     - name: commandDir
-      value: "../../../../commands"
+      value: "../../../../../lib/commands"
     - name: gkoNamespace
       value: default
     - name: exportedApiYaml
