+++ test/e2e/chainsaw/tests/apis/notifications/updatingGroupRefs/chainsaw-test.yaml
@@ -38,7 +38,7 @@ spec:
     - name: apiName
       value: (join('-', [$apiVersion, $test.metadata.name]))
     - name: commandDir
-      value: "../../../../commands"
+      value: "../../../../../lib/commands"
     - name: gkoNamespace
       value: default
     - name: exportedApiYaml
