+++ test/e2e/chainsaw/tests/apis/state/startStopApi/chainsaw-test.yaml
@@ -37,7 +37,7 @@ spec:
     - name: apiName
       value: (join('-', [$apiVersion, $test.metadata.name]))
     - name: commandDir
-      value: "../../../../commands"
+      value: "../../../../../lib/commands"
     - name: gkoNamespace
       value: default
     - name: exportedApiYaml
