+++ test/e2e/lib/commands/getApiNotifications.mjs
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-import { mapiClient } from '../lib/mapiClient.mjs';
+import { mapiClient } from '../gravitee/mapi/client.mjs';
 
 const apiId = argv.api_id;
 const orgId = argv.org_id || 'DEFAULT';
